move pipe read limit to gint64
We had a mix of size_t and gint64. Just use gint64 everywhere.

+++ libvips/include/vips/connection.h
@@ -89,7 +89,7 @@ GType vips_connection_get_type( void );
 const char *vips_connection_filename( VipsConnection *connection );
 const char *vips_connection_nick( VipsConnection *connection );
 
-void vips_pipe_read_limit_set( size_t limit );
+void vips_pipe_read_limit_set( gint64 limit );
 
 #define VIPS_TYPE_SOURCE (vips_source_get_type())
 #define VIPS_SOURCE( obj ) \
